+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.3.1] - December 30, 2020 (Kent, CT)
+### Fixed
+- Incorrect output of ``pofp.partition_data_by_break_points`` caused by unsorted data from ``trees.rolling_search``.
 
 ## [v0.3.0] - December 30, 2020 (Kent, CT)
 ### Added