+++ CHANGELOG.md
@@ -3,9 +3,14 @@ All important changes to the decitala package will be documented here.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.3.1] - December 30, 2020 (Kent, CT)
+## [v0.3.1] - December 31, 2020 (Kent, CT)
+### Added
+- Documentation page for ``vis.create_tree_diagram`` (Fragment Tree visualization using Treant.js.)
+- Log message telling the user if a database has already been made (useful for working in Jupyter).
+
 ### Fixed
 - Incorrect output of ``pofp.partition_data_by_break_points`` caused by unsorted data from ``trees.rolling_search``.
+- Documentation typo for ``utils.filter_single_anga_class_fragments``.
 
 ## [v0.3.0] - December 30, 2020 (Kent, CT)
 ### Added